Adding in a "develop" command in the cli
Adding a develop command which spawns a file watcher, ands runs the incremental build whenever files in src or content are changed. Opted for a new command, as it can be swapped when better ways to do this arrive

Currently breaks if there's an error that hits unwrap. Started diving into those, got very scared, backed right on out and will approach in another pr