Enable *not* prerendering `SetDataForSlug`s
Currently if you `setDataForSlug` it automatically gets
recognized as a "page" (page is not a real concept), and gets
rendered to html as a result. This commit adds the ability to set
`prerender: false`, indicating that the structure should not
be turned into html.

This enables various scenarios, for example:

* Store code in remote CMSs, such as additional MDX components, then
  import and use that code for prerendering statically or in the browser
  with a dynamic import.
* imports from "outside src/", such as when you have toast-tools/mdx
  set on a directory, and you also have additional .js files being
  imported into those MDX files.

  toast-tools/mdx can now scan and automatically process the imports
  from mdx -> .js, using SetDataForSlug on non-prerendered .js files.